fix(app): try requesting passkey using window.location.hostname
hopefully fixes iOS Safari crashing after latest passkey update
